LU-7646 lnet: Stop Infinite CON RACE Condition
In current code, when a CON RACE occurs, the passive side will
let the node with the higher NID value win the race.

We have a field case where a node can have a "stuck"
connection which never goes away and is the trigger of a
never-ending loop of re-connections.

This patch introduces a counter to how many times a
connection in a connecting state has been the cause of a CON RACE
rejection. After 2 times, we assume the connection is stuck
and let the other side (with lower NID) win.

Note: this patch is meant to bring parity between the
master branch and an equivaltent fix landed to other repos.
A better patch may follow under this Jira ticket later on.

Signed-off-by: Doug Oucharek <doug.s.oucharek@intel.com>
Change-Id: I32e035806e95868b13c28c42e241b969940a35c9