LU-7689 obdclass: limit lu_site hash table size on clients
Allocating a big hash table using the formula for osd
does not really work for clients. We will create new
hash table for each mount on a single client which is
a lot of memory more than expected.

This patch limits the hash table up to 8M on clients,
which has 524288 entries.

Signed-off-by: Li Dongyang <dongyang.li@anu.edu.au>
Change-Id: I908fda102ec5fd46c1325e0e41f5fe291aaa3378