LU-7700 build: Fix autoreconf by using AC_CONFIG_MACRO_DIRS
Autoreconf does not work because we are not letting autotools
know about all of the various m4 macro directories in Lustre's
tree anywhere other than aclocal's command line in autogen.sh

As of automake 1.13 at least, AC_CONFIG_MACRO_DIRS exists and
can both replace our use of aclocal command line includes in
autogen.sh, and express the includs in a way that allows
autoreconf to work.

The only problem, of course, is that automake 1.13 is relatively
new (2012), and is not currently in some (all?) RHEL6 releases.

Bump the required automake version to 1.13 in AM_INIT_AUTOMAKE
to reflect that requirement.

Change-Id: I1594aade093115da3181e2dd5a99c81176ee7829
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>