LU-5134 utils: Add parallel option to lctl set_param
Add a "-p" option to lctl set_param to enable setting multiple matched
parameters in parallel. When called with "-p", lctl will set up a work
queue of matched file names and spawn a fixed number of threads per
CPU. Each thread will pull items off the work queue, write to the file
associated with each work item, and return when there are no more
items on the work queue.

A bit field called po_parallel is added to the struct param_opts to
indicate whether the set_param should be executed in parallel. If in
parallel, jt_lcfg_setparam initializes a work queue and passes it to
do_param_op, which adds each matched item to the work queue. Once
jt_lcfg_setparam has called do_param_op for each param-value pair, it
passes the work queue to sp_run_threads, which creates threads, each
of which call write_param to set the parameter. If not in parallel,
jt_lcfg_setparam does not pass a work queue to do_param_op, and
do_param_op directly calls write_param on each matched param.

param_display was renamed to do_param_op to more accurately reflect
what it does, and it is defined as a macro to allow callers to call it
with 5 arguments whether we have pthread support or not.

If lctl is compiled without pthread support, "lctl set_param" will
still accept the "-p" option, but it will print a warning message, and
it will set the parameters in series.

Sanity test 124d was added to test the "-p" option to lctl set_param.
It also checks that writing "clear" to lru_size clears all unused
locks. There was no test for this before. This also involved creating
total_unused_locks and total_used_locks functions in test-framework.sh
so that all tests could perform these operations consistently.

Fixed minor style issue pointed out in an LU-5030 patch.

Signed-off-by: Ryan Haasken <haasken@cray.com>
Change-Id: I3f96a6f06c50d4ba2ce97050c35f46b976dfc005