LU-7846 nodemap: add fileset info to nodemap
Give the ability to set a fileset information on a nodemap.
When set, a client pertaining to this nodemap will be allowed to
mount only the subdirectory represented by this fileset.

Fileset info of nodemap is stored permanently by using
'lctl set_param -P' commands. It means MDS and OSS will be able to
retrieve fileset info of nodemap after a restart (when they read
params llog).
MGS will not see fileset info after restart, but this does not prevent
the feature from working.

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Change-Id: Ic16abaeb0b2fe583a9a4745258a82bbd23fc259e