LU-7613 Modified ll_find_alias to avoid cache corruption
ll_find_alias is responsible for getting alias for inode
which can be reused. Directories are asumed to have unique
alias. Where in case of non-directories there can be multiple
aliases. In case of lustre there can be two type of aliases
i.e. discon_alias and invalid_alias. Usage of discon_alias in
case of non-directories may corrupt dcache and leads to kernel
crash.Changes made to avoid use of discon_alias in case of
non-directories.

Seagate-bug-id: MRP-2739
Change-Id: Ieb9dabab0784bdb3c52e2cb32be1a766ffe55313
Signed-off-by: Lokesh Nagappa Jaliminche <lokesh.jaliminche@seagate.com>