LU-7144 tests: print client/server versions for tests
Print the client and server versions for sanity-scrub and
sanity-lfsck so that we can see what they are reporting.

This is relevant for testing the upstream kernel version,
since it appears to be trying to run sanity-scrub and
sanity-lfsck when it should be skipping them instead, but
it is useful information to have in general.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I7b4c5d273883dfc18ab20b43176620842f3ebbe5