LU-7429 test: LFSCK isn't getting completed within given time.
The test 23c of sanity-lfsck.sh times out after the specified time in
wait_update not letting the LFSCK to recover the dangling name entry.
This causes the test to fail and it is evident from the logs that the
LFSCK was under progress. Just added an extra parameter into the
wait_update function that will accept the interval for checking the
specified condition for size of the file and also increased the time
for LFSCK to let it recover the dangling name entry.

Change-Id: I2ccb8cfae03e84ce68ec9c7a142816c1b2bab8ee
Seagate-bug-id: MRP-3134
Signed-off-by: Kirtankumar Krishna Shetty <kirtan.shetty@seagate.com>