LU-7656 tests: tar fix for replay-single/70c
The tar process runs in an infinite loop, and the removal/cleanup
of files interferes in the process and causes tar to fail.

The code flow could take two paths:
1. Mdt recovery has completed, catch exit signal.
In this case, put a check and confirm if the mdt recovery has
completed, if yes, then exit and perform cleanup
commands (remove files).

2. Main process code flow.
Wait for tar process to complete and then perform cleanup commands.

Also, instead of removing files within the tar process, keep the
removal of files in the cleanup part.

Seagate-bug-id: MRP-2925
Signed-off-by: Noopur Maheshwari <noopur.maheshwari@seagate.com>
Change-Id: Ia1c840dde554a2f1487aa9aebb70e5dcea29b806
Reviewed-on: http://review.whamcloud.com/18732
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ujjwal Lanjewar <ujjwal@yahoo.com>
Reviewed-by: Wei Liu <wei3.liu@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>