LU-7867 osd-ldiskfs: compatible check for IDIF OI consistency
For old IDIF, the OST index is not part of the IDIF, means the IDIF in
LMA EA does not contain the OST index, but the IDIF used for locating
the OST-object may contains OST index, such difference would misguide
the OSD to reguard the IDIF OI mapping invalid as to trigger OI scrub
unexpectedly. The right solution is to make some compatible check to
distinguish whether the invalid IDIF OI mapping is caused by OST index
issue or not to avoid triggered OI scrub unnecessarily.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I1fba55064396850bc9634c6689d38f4683c3b75e