LU-7813 lod: inherit pool from parent if unspecified
If the OST pool is not set in a specific file layout, but the pool
is set on the parent directory, inherit the pool from the parent.

This makes OST pools much more usable, since the expectation is that
files created in a directory with a pool set will be part of the pool,
even if there was a need to set explicit striping for the file.

Test-Parameters: fortestonly
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I4747ec3aa829fa9beec50e4c85c5238b2f4f5ff7