LU-7643 build: Remove Linux version string from RPM release field
With the landing of the patch:

  "LU-5614 build: use %kernel_module_package in rpm spec"

we neither need, nor desire, to tie a lustre build to a specific
Linux kernel version.  Instead we want to allow the weak updates
system to symlink the lustre modules into any an all kernel
directories that are ABI compatible.

Therefore, we remove the Linux version string from Lustre's
RPM release field.

Note that "%define krequires" is moved to a SLES-only section
when the 0%{?build_src_rpm} conditional is removed.  This may
seem like an odd place to put it at first glance.  But the
next line is the only remaining consumer of krequires, so
we are just moving it to the last place that it might be needed.
This could perhaps be simplified further in a future commit.

When removing the no longer needed %{fullrelease} from the Release
field, we replace it with a 1 and with the %{?dist} string.
%{?dist} is customary for Fedora, RHEL, and likely most rpm-based
distros.

Change-Id: I8306f48d5bf0dffcae2888505ec7f0f63e47fc08
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>