LU-5960 lnet: add in functionality of lctl to DLC
One of the barriers to moving from lctl to lnetctl
is the lack of certain functionality. This patch
adds some of the feature that are missing back.

The following commands have been added:
. show peers
. add peer
. del peer
. show connections
. close connection
. push connection

Besides adding in the missing pieces I have added
new features such as filtering the listing of peer
or connections based on properties such as NID,PID
or even LNet type (e.g o2ib1) to the lnetconfig
library. From the lnetctl command line one can
call lnetctl peer show --nid 10.0.0.1@tcp and it will
only show that specific peer. Listing of a grouping
of peers is also possible with for example lnetctl
peer show --nid o2ib12 which will display all peers
belong to the o2ib12 net configuration. Lastly the
ability to use DNS hostnames was also added.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Change-Id: Id102e8a6148d9e2ef8c88880df86e237a14c5ebb