LU-7759 llite: handle inactive OSTs better in statfs
Enable the lazystatfs mount option by default, so that "df" does not
hang when an OST is temporarily offline.

Change the order of checks for inactive OSCs in lov_prep_statfs_set()
so that administratively disabled OSTs do not generate any output in
"lfs df" at all, to avoid needlessly cluttering the output.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I993761a7eb120a36a1b80c2822cb6d8011ccab07