LU-7981 llite: fix double read of lli_trunc_sem
The lli_trunc_sem is taken in 'read' mode in both
ll_page_mkwrite and vvp_io_fault_start. This can lead to a
deadlock with another thread which asks for the semaphore
in write mode between thse two read calls.

Use cl_io_is_mkwrite in the vvp layer to only take the
lli_trunc_sem when we are not doing a mkwrite call.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I7341d179f550835ccebfb1bb2d90260c1d5d0ff8