LU-7988 hsm: execute HSM coordinator once per second at most
When there is heavy HSM usage, each new HSM request can trigger the
HSM coordinator, which may run many times per seconds. When it is
running it locks the HSM catalog (cdt_llog_lock) preventing any other
HSM operation to happen, such as insertion, removal or dumping of the
requests.

Limit the coordinator to run once per second, and only if there is
work to do. It will still execute the loop once every 10 seconds (or
as defined by the loop_period parameter) to do housekeeping.

Signed-off-by: frank zago <fzago@cray.com>
Change-Id: Ide3f061f8943a3088ea713993521897fb74e5d99