LU-5620 ptlrpc: Add QoS for opcode in NRS-TBF
This patch add a new QoS feature in TBF policy which could
limits the rate based on opcode.

The syntax is like:
    lctl set_param x.x.x.nrs_tbf_rule=
         "[reg|hp] start <rule_name> <arguments>..."
Start the tbf opcode QoS:
    lctl set_param ost.OSS.ost_io.nrs_policies="tbf opcode"
Limit the ost_read and ost_write respectively:
    lctl set_param ost.OSS.ost_io.nrs_tbf_rule=
	 "start ost_r {ost_read} 100"
    lctl set_param ost.OSS.ost_io.nrs_tbf_rule=
	 "start ost_w {ost_write} 200"
Limit both ost_read and ost_write:
    lctl set_param ost.OSS.ost_io.nrs_tbf_rule=
         "start ost_rw {ost_read ost_write} 200"

The limit numbers like 100, 200 mean the number of
requests per second.

And, the opcode-based policy can not be combined with
NID-based and JobID-based policies now.

Test-Parameters: alwaysuploadlogs
Signed-off-by: Wu Libin <lwu@ddn.com>
Signed-off-by: Wang Shilong <wshilong@ddn.com>
Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: I4ff93972df560ad1ebc8e38e942d503518a835c7