LU-7988 hsm: improve hsd.request reset
There is no need to reset hsd.request since hsd.request is already
reset in a for loop above. This saves a memset of 16 bytes per
request, each time the arbitor runs.

The request's hal_sz also doesn't need to be reset to 0 since it is
only used when hal exists.

Test-Parameters: trivial testlist=sanity-hsm
Signed-off-by: frank zago <fzago@cray.com>
Change-Id: I2b1296ecf6469990983cf25d40a8c97205344525