LU-7988 hsm: only browse entries of hsd.request when needed
In the coordinator callback, only the entries in hsd.request from 0 to
hsd.request_cnt are used. So use that property instead of walking the
whole array. Consequently, the next entry available in hsd.request is
always at the hsd.request_cnt position, so use it instead of trying to
find an empty slot.

Test-Parameters: forbuildonly
Signed-off-by: frank zago <fzago@cray.com>
Change-Id: I2be0fe5ce918ded028bb260ef345a859b2cc41d4