LU-7988 hsm: defer all new archive hals
All new archive HALs are enqueued on a coordinator list, and the
coordinator will process them once every "wait_event_time"
seconds. This reduces the number of times hsm_find_compatible and thus
cdt_llog_process are called.

Other types of HAI (restore, cancel, ...) are processed as usual. When
one of these is received, the deferred list of archive HAI, if not
empty, is processed first.

Test-Parameters: forbuildonly
Signed-off-by: frank zago <fzago@cray.com>
Change-Id: Ie4a5eb0c1115dad7b843e599a01042db3b8911b7