LU-7110 tests: sanity-hsm.sh can erase all files on filesystem
For SINGLEAGT=client, $HSM_ARCHIVE & $hsm_root is empty.
Thus purging of archive in copytool_setup() will result
in execution of "rm -rf /*", resulting in permanent loss
of files as well as cluster node.

This change fixes it by checking the respective variables
like HSM_ARCHIVE or hsm_root to be non-empty.

Change-Id: I73cdc08a58409ae1f3ba85a9d12077062a38b321
Seagate-Bug-Id: MRP-2659
Signed-off-by: Parinay Kondekar <parinay.kondekar@seagate.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/6471
Tested-by:Jenkins
Reviewed-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>
Reviewed-by: Alexander Lezhoev <alexander.lezhoev@seagate.com>
Reviewed-on: http://review.whamcloud.com/16291
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: jacques-Charles Lafoucriere <jacques-charles.lafoucriere@cea.fr>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>