LU-7887 build: revert of use "make rpms" for building Lustre RPMs
Testing the theory that recent lustre-init fails of all sles builds
is really a build problem introduced by the lbuild change.
If the theory proves true will need a fix in current lbuild.

This revert is only for testing purposes, not to be landed.

Test-Parameters: mdsdistro=sles11sp4 ossdistro=sles11sp4 \
  clientdistro=sles11sp4 mdsfilesystemtype=ldiskfs \
  mdtfilesystemtype=ldiskfs ostfilesystemtype=ldiskfs \
  testgroup=review-ldiskfs fortestonly

Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Change-Id: Icc88852fdea4c7195e69ce977b09170fe61c1412