LU-8014 hsm: remove kuc_ispayload
That function takes a pointer, decrements it and dereference the new
address. However we have no idea if the page it's in is
readable or even exists. Such dereference would cause an oops.

That function is only called from mdt_hsm_agent_send, which has itself
2 callers. In the most common case, from mdt_coordinator(), the buffer
passed is already a KUC, but in the less used case, from
hsm_cancel_all_actions(), it is not. Allocating the KUC buffer in
mdt_coordinator() doesn't buy us anything. So remove that code from
the coordinator, which means mdt_hsm_agent_send() is now always called
with a regular buffer, and kuc_ispayload can go.

Signed-off-by: frank zago <fzago@cray.com>
Change-Id: I99c29d4223c14b6fb778215032cd07b6e4676990