LU-7915 mdc: add a tunable to enable opencache
Lustre has the ability to cache opens on the client with a DLM
openlock, but this isn't done for regular opens by default because
it has extra overhead compared to uncached opens, but only for NFS
opens because the knfsd repeatedly opens the same file.

It would be worthwhile to implement a tunable to enable opencache
on a per-client basis and then measure the performance impact of
this tunable for normal usage.

Signed-off-by: Emoly Liu <emoly.liu@intel.com>
Change-Id: I8226c16adb46e374b84604716bbd1e777b9f8a7e