LU-5092 nodemap: users of ted_nodemap should take ref
In order to quickly map clients, the client's assigned nodemap is
saved on the export data struct. When mapping occurs, a reference to
that nodemap is used for mapping. If no reference is taken during
mapping and the nodemap is put elsewhere, then the nodemap could be
destroyed before the mapping is finished. This patch ensures that a
reference is taken before mapping.

Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Change-Id: I5c4d707970b1a86fb061a501240182ff3dc7be51