LU-8051 tests: Skipping sanity tests 300[a-d] for old servers.
The tests 300[a-i] were added in version 2.7.0 for the newer
version of the code to test striped directory related features
and hence the tests 300[a,b,c,d] used to fail for the v2.7.0
and above client with older server versions. Added skip
statement for the tests if server is of an older versions.

Seagate-bug-id: MRP-3393
Change-Id: I26f271c1a4aa62c8f41c391e0c595f1ccaf18f8b
Signed-off-by: Kirtankumar Krishna Shetty <kirtan.shetty@seagate.com>