LU-8010 mdt: unlock layout when failing restore
Previously to this patch layout was not unlocked when a restore
was failed before being sent to CT, leading to a situation where
other requestors hang and also to an orphan restore_handle to be
kept on CDT's list of registered restore actions.
Only way to clear situation then, was to stop CDT.
This situation could at least occur if a restore was canceled
but the CT does not handle cancel operation, allowing the restore
to complete but also to have new restore requests to be
registered in the mean time and then to be failed due to
incompatible (!released) file state.

Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>
Change-Id: Ib1ba9156793a230d256ff80d74372813f10b0321