LU-7699 build: Convert version underscores to dashes for dpkg
For rpm, we can't have dashes in the version string so we use
underscores.  For dpkg, the requirements are reversed.

In the "debs" make target, we convert underscores in the version
string to dashes.  We also trim down the comment that no longer
really makes sense with the new versioning scheme.

Change-Id: I6d600fc4a16be60c664ef0662b482c31877fbfea
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>