LU-8055 ldev: Further improvements to ldev.
Allow for the hostname and backend file system to be used in command
substitution. Output can now be filtered by using -R <role> or
--role <role> as command line parameters, where role is a string such
as "mdt", "ost", "mgs", or "mgt". All labels for a file system can be
printed by using the -F <name> or --filesys <name> as
command line parameters, where name is the file system of interest.

Signed-off-by: Giuseppe Di Natale <dinatale2@llnl.gov>
Change-Id: I50726371dce809e4d1b4f3de6f530d6f637856a1