LU-8062 test: fix fail_val in recovery-small/115b
The fail_loc OBD_FAIL_OST_ENOSPC is caught in 2 places:
1. ofd_statfs() - where it is checked against the fail_val
and if fail_val matches the OST ID it is caught here
itself and hence test fails.
2. tgt_brw_write() - the actual place where the fail_loc should
be caught.

The patch makes the fail_loc to be caught at the appropriate place
by setting the fail_val to $OSTCOUNT. So even if fail_loc is set
in the statfs part it is not caught there and therefore caught in
tgt_brw_write().

Seagate-bug-id: MRP-3150
Change-Id: I88cb713bd8de52de79937d338fb2533d99c1ee0f
Signed-off-by: Bhagyesh Dudhediya <bhagyesh.dudhediya@seagate.com>