LU-7592 osd-ldiskfs: remove force_over_128 warning
Warning message that advice to remove force_over_128 option
is shown after 3.0.53 version.

It would actually be better to just stop supporting the
force_over_128tb option completely when this version is
hit, instead of adding the warning, since that will cause
build failures if it is ever hit. Then, this code can be
deleted at leisure after that point instead of having
to be deleted before 3.0.53 is hit.

Signed-off-by: Artem Blagodarenko <artem.blagodarenko@seagate.com>
Change-Id: I73c0f4a62b6a80bde26fcb4eb10cf7a9ff60ee88