LU-8041 mdd: increasing only atime update on close
The atime update on close should be increasing only, otherwise,
the atime could be overwritten by stale client's atime.

This patch also changed client to clear out the ATTR_ATIME
flag fro close RPC when atime == 0, so that patch client won't
overwrite the unpatched server with 0 atime.

A test case sanity 39p is added to verify that close won't
overwrite atime with zero, and sanity 39l is adjusted as well.

Signed-off-by: Niu Yawei <yawei.niu@intel.com>
Change-Id: I40d1103ddfea184d82170a6795f89ce303a5c14e