LU-6215 o2iblnd: port to new fast reg API introduced in 4.4
Remove the allocation of fastreg page list, as the page
vector is now private to the provider. Just pass tx_frags
to ib_map_mr_sg() and construct ib_reg_wr.
Defer the conversion of tx_frags to tx_pages and only do
it when the new API is not available.

Linux-commit: 4c67e2bfc8b7121d51434362fa7c2d012f8bcf1b
Linux-commit: 39bfc271bd687be2c8e396e976c0fb9a97963400

Signed-off-by: Li Dongyang <dongyang.li@anu.edu.au>
Change-Id: I5c62b0370ad6ddcc93102a29343491968e4446d0