LU-8113 build: Remove obsolete, unused meta packages
The meta packages, deps-rhel and deps-sles, were introduced
nine years ago and never used.  This removes those dead lines
in the configuration files.

We would not want to use this method in the future any way.
Better approaches would either be to use multiple spec files
or to use distro-specific macros and conditionals for expressing
dependencies.

Test-Parameters: trivial
Change-Id: I9ad7195e0504f638bca84984af499d959c37adb3
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>