LU-8116 build: Cleanup GSS configure script messages
Some of the GSS configure warning are unnecessarily multi-line,
and also redundant.  For instance:

  checking for keyctl_search in -lkeyutils... no
  configure: WARNING:

  libkeyutils is not found, which is required by gss keyring backend

  configure: WARNING:

  Cannot enable gss keyring. See above for details.

Why so many lines?  Why the double warning?

Also, there is a bug in the "for Kerberos v5" check where it
fails to use AC_MSG_RESULT in the negative case.  A completely
different function winds up appending an AC_MSG_WARN:

  checking for Kerberos v5... configure: WARNING: not found!

This patch addresses those minor issues.

Change-Id: Ic9fb8b5687849688d965cc67b394e4eb569204be
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>