LU-7659 mdc: expose hsm requests through char devices
Register one character device per filesystem to allow non-llapi to
read them and to make delivery more efficient.

HSM requests (actually "struct hsm_action_list") can be published
from the kernel with the "ct_cdev_hal_publish" function.

Opening the char device in reading mode is similar to subscribing to
incoming requests. Reading from the char device will copy requests
published after the open in the order they were published. Requests
published prior to the open operation are simply not visible. The
reading offset is used to provide the archive_mask of the copytool
and only retrieve the right requests (the other requests are ignored)

Requests are stored in a linked list and are not dropped until every
reader who can read it actually does. The close operation releases
allocated resources. Requests are dropped accordingly.

An ioctl is implemented to set the archive_mask you want to use to
filter HSM requests (CT_CDEV_IOC_SET_ARCHIVE_MASK). It takes the
archive_mask as its 3rd argument. The archive_mask defaults to 0
which is "broadcast".

(Writing in the char device fails with errno set to EOPNOTSUPP)

Signed-off-by: Quentin Bouget <quentin.bouget.ocre@cea.fr>
Change-Id: I32091aa6221e48e2fc2cc0bcc3cb3d013df0b663