LU-3606 fallocate: Implement fallocate(2) support for Lustre
Phase-1:
This patch adds fallocate(2) support for Lustre. fallocate(2) method
of the inode_operations/file_operations is implemented and
transported to the OSTs to interface with the underlying OSD's
fallocate(2) code.

A new RPC, OST_PREALLOC has been added for space preallocation.
The fallocate functionality (prealloc) in CLIO has been multiplexed
with CIT_SETATTR.

The Phase-2 of this patch will add "punch" functionality of
fallocate(2).

Signed-off-by: Swapnil Pimpale <spimpale@ddn.com>
Signed-off-by: Li Xi <lixi@ddn.com>
Change-Id: I03f27d356616fbf3a3ab8e6309af26c00434d81b