LU-7352 tests: don't fail conf-sanity test_78 on ENOSPC
Add debugging to conf-sanity.sh test_78 to see why this test is
regularly reporting ENOSPC creating 100x 1MB test files, when
there should be many GB of space available during the test.

Don't fail the test if the OST runs out of space, or if the MDS
refuses to create more files when OST has no available blocks,
since the main goal of this test is to verify ldiskfs resizing.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Id5a7049aa434c794a5a8f8aa5398bf7ca9bcab07